refactor: move evaluation code to app/utils/evals.py
